tools/probe: Split the demuxer and offline command to separate files
