+++ tools/probes/CMakeLists.txt
@@ -3,8 +3,8 @@
 cmake_minimum_required(VERSION 3.13)
 
 add_executable(sof-probes
+	probes_demux.c
 	probes_main.c
-	../../src/math/numbers.c
 )
 
 target_compile_options(sof-probes PRIVATE
